requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-bigquery==3.25.0
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic==2.9.2
